appointment follow up as structured question
